Global setup/teardown Remove serial runner
