+++ package.json
@@ -231,7 +231,6 @@
     "@babel/cli": "^7.21.5",
     "@babel/preset-typescript": "^7.21.4",
     "@faker-js/faker": "^8.0.2",
-    "@getoutline/jest-runner-serial": "^2.0.0",
     "@relative-ci/agent": "^4.1.3",
     "@types/addressparser": "^1.0.1",
     "@types/body-scroll-lock": "^3.1.0",
