+++ server/commands/accountProvisioner.test.ts
@@ -5,14 +5,8 @@ import { TeamDomain } from "@server/models";
 import Collection from "@server/models/Collection";
 import UserAuthentication from "@server/models/UserAuthentication";
 import { buildUser, buildTeam, buildAdmin } from "@server/test/factories";
-import {
-  setupTestDatabase,
-  setCloudHosted,
-  setSelfHosted,
-} from "@server/test/support";
+import { setCloudHosted, setSelfHosted } from "@server/test/support";
 import accountProvisioner from "./accountProvisioner";
-
-setupTestDatabase();
 
 describe("accountProvisioner", () => {
   const ip = "127.0.0.1";
