+++ server/commands/documentPermanentDeleter.test.ts
@@ -2,12 +2,9 @@ import { subDays } from "date-fns";
 import { Attachment, Document } from "@server/models";
 import DeleteAttachmentTask from "@server/queues/tasks/DeleteAttachmentTask";
 import { buildAttachment, buildDocument } from "@server/test/factories";
-import { setupTestDatabase } from "@server/test/support";
 import documentPermanentDeleter from "./documentPermanentDeleter";
 
 jest.mock("@server/queues/tasks/DeleteAttachmentTask");
-
-setupTestDatabase();
 
 describe("documentPermanentDeleter", () => {
   it("should destroy documents", async () => {
