+++ server/commands/starUpdater.test.ts
@@ -1,9 +1,7 @@
 import { Star } from "@server/models";
 import { buildDocument, buildUser } from "@server/test/factories";
-import { findLatestEvent, setupTestDatabase } from "@server/test/support";
+import { findLatestEvent } from "@server/test/support";
 import starUpdater from "./starUpdater";
-
-setupTestDatabase();
 
 describe("starUpdater", () => {
   const ip = "127.0.0.1";
