+++ server/commands/subscriptionCreator.test.ts
@@ -1,11 +1,8 @@
 import { Subscription, Event } from "@server/models";
 import { sequelize } from "@server/storage/database";
 import { buildDocument, buildUser } from "@server/test/factories";
-import { setupTestDatabase } from "@server/test/support";
 import subscriptionCreator from "./subscriptionCreator";
 import subscriptionDestroyer from "./subscriptionDestroyer";
-
-setupTestDatabase();
 
 describe("subscriptionCreator", () => {
   const ip = "127.0.0.1";
