+++ server/commands/teamPermanentDeleter.test.ts
@@ -6,10 +6,7 @@ import {
   buildTeam,
   buildDocument,
 } from "@server/test/factories";
-import { setupTestDatabase } from "@server/test/support";
 import teamPermanentDeleter from "./teamPermanentDeleter";
-
-setupTestDatabase();
 
 describe("teamPermanentDeleter", () => {
   it("should destroy related data", async () => {
