+++ server/commands/userDemoter.test.ts
@@ -1,10 +1,7 @@
 import { CollectionPermission, UserRole } from "@shared/types";
 import { CollectionUser } from "@server/models";
 import { buildUser, buildAdmin, buildCollection } from "@server/test/factories";
-import { setupTestDatabase } from "@server/test/support";
 import userDemoter from "./userDemoter";
-
-setupTestDatabase();
 
 describe("userDemoter", () => {
   const ip = "127.0.0.1";
