+++ server/policies/team.test.ts
@@ -1,12 +1,6 @@
 import { buildUser, buildTeam, buildAdmin } from "@server/test/factories";
-import {
-  setCloudHosted,
-  setSelfHosted,
-  setupTestDatabase,
-} from "@server/test/support";
+import { setCloudHosted, setSelfHosted } from "@server/test/support";
 import { serialize } from "./index";
-
-setupTestDatabase();
 
 it("should allow reading only", async () => {
   await setSelfHosted();
