+++ server/queues/processors/BacklinksProcessor.test.ts
@@ -1,11 +1,8 @@
 import { Backlink } from "@server/models";
 import { buildDocument } from "@server/test/factories";
-import { setupTestDatabase } from "@server/test/support";
 import BacklinksProcessor from "./BacklinksProcessor";
 
 const ip = "127.0.0.1";
-
-setupTestDatabase();
 
 describe("documents.publish", () => {
   test("should create new backlink records", async () => {
