+++ server/queues/tasks/InviteReminderTask.test.ts
@@ -1,10 +1,7 @@
 import { subDays } from "date-fns";
 import InviteReminderEmail from "@server/emails/templates/InviteReminderEmail";
 import { buildInvite } from "@server/test/factories";
-import { setupTestDatabase } from "@server/test/support";
 import InviteReminderTask from "./InviteReminderTask";
-
-setupTestDatabase();
 
 describe("InviteReminderTask", () => {
   it("should not destroy documents not deleted", async () => {
