+++ server/queues/tasks/index.ts
@@ -1,10 +1,12 @@
 import path from "path";
 import { glob } from "glob";
+import env from "@server/env";
 import Logger from "@server/logging/Logger";
 import { requireDirectory } from "@server/utils/fs";
 import BaseTask from "./BaseTask";
 
 const tasks = {};
+const rootDir = env.ENVIRONMENT === "test" ? "" : "build";
 
 requireDirectory<{ default: BaseTask<any> }>(__dirname).forEach(
   ([module, id]) => {
@@ -16,7 +18,7 @@ requireDirectory<{ default: BaseTask<any> }>(__dirname).forEach(
 );
 
 glob
-  .sync("build/plugins/*/server/tasks/!(*.test).js")
+  .sync(path.join(rootDir, "plugins/*/server/tasks/!(*.test).[jt]s"))
   .forEach((filePath: string) => {
     // eslint-disable-next-line @typescript-eslint/no-var-requires
     const task = require(path.join(process.cwd(), filePath)).default;
