+++ server/test/globalSetup.ts
@@ -1,0 +1,14 @@
+import { sequelize } from "@server/storage/database";
+
+module.exports = async function () {
+  const sql = sequelize.getQueryInterface();
+  const tables = Object.keys(sequelize.models).map((model) => {
+    const n = sequelize.models[model].getTableName();
+    return (sql.queryGenerator as any).quoteTable(
+      typeof n === "string" ? n : n.tableName
+    );
+  });
+  const flushQuery = `TRUNCATE ${tables.join(", ")} CASCADE`;
+
+  await sequelize.query(flushQuery);
+};
