+++ server/test/globalTeardown.ts
@@ -1,0 +1,5 @@
+import { sequelize } from "@server/storage/database";
+
+module.exports = async function () {
+  await sequelize.close();
+};
